Add files via upload
Class with the method: double static getNumInput(String param) which allows getting inputs through a dialog box for each orbital parameter (defined with string in the input). Need to copy-paste code from this main method to point before the orbit is defined to make it possible to choose orbit before displaying. Correct parameters need to be entered for LEO, SSO, and GEO options.